fix spend: check duplicate recipients addresses
This a temporary fix to show an error message when
two recipients addresses are the same and prevent
user to go next step. Because the recipients outputs
are stored in a address-amount map, a recipient output
with the same address than an other replace it.
In order to change this, revaultd needs to accept a
vector of (address, amount) tuple instead of a map.